+++ hmkit-fleet/src/main/kotlin/network/TelematicsRequests.kt
@@ -65,6 +65,37 @@ internal class TelematicsRequests(
         return postCommand(encryptedCommand, accessCertificate)
     }
 
+    suspend fun sendCommandV05(
+        command: Bytes,
+        accessCertificate: AccessCertificate
+    ): Response<Bytes> {
+        val nonce = getNonce()
+
+        if (nonce.error != null) return Response(null, nonce.error)
+
+        val encryptedCommand =
+            crypto.createTelematicsContainer(
+                command,
+                privateKey,
+                certificate.serial,
+                accessCertificate,
+                Bytes(nonce.response!!)
+            )
+
+        val encryptedCommandResponse = postCommandV05(encryptedCommand, accessCertificate)
+
+        if (encryptedCommandResponse.error != null) return encryptedCommandResponse
+
+        val decryptedResponseCommand = crypto.getPayloadFromTelematicsContainer(
+            encryptedCommandResponse.response!!,
+            privateKey,
+            accessCertificate,
+        )
+
+        return Response(decryptedResponseCommand)
+    }
+
+
     private suspend fun getNonce(): Response<String> {
         val request = Request.Builder()
             .url("${baseUrl}/nonces")
@@ -83,7 +114,7 @@ internal class TelematicsRequests(
         val response = call.await()
 
         return tryParseResponse(response, HttpURLConnection.HTTP_CREATED) { body ->
-            val jsonResponse = Json.parseToJsonElement(body) as JsonObject
+            val jsonResponse = jsonIg.parseToJsonElement(body) as JsonObject
             val nonce = jsonResponse.jsonObject["nonce"]?.jsonPrimitive?.content
             Response(nonce, null)
         }
@@ -114,7 +145,7 @@ internal class TelematicsRequests(
 
         val responseObject = try {
             if (response.code == 200 || response.code == 400 || response.code == 404 || response.code == 408) {
-                val telematicsResponse = Json.decodeFromString<TelematicsCommandResponse>(responseBody)
+                val telematicsResponse = jsonIg.decodeFromString<TelematicsCommandResponse>(responseBody)
 
                 // Server only returns encrypted data if status is OK
                 val decryptedData = if (telematicsResponse.status == TelematicsCommandResponse.Status.OK) {
@@ -137,12 +168,43 @@ internal class TelematicsRequests(
             } else {
                 // try to parse the normal server error format.
                 // it will throw and will be caught if server returned unknown format
-                TelematicsResponse(errors = Json.decodeFromString(responseBody))
+                TelematicsResponse(errors = jsonIg.decodeFromString(responseBody))
             }
         } catch (e: Exception) {
             TelematicsResponse(errors = listOf(Error(title = "Unknown server response", detail = e.message)))
         }
 
         return responseObject
+    }
+
+    private suspend fun postCommandV05(
+        encryptedCommand: Bytes,
+        accessCertificate: AccessCertificate,
+    ): Response<Bytes> {
+        val request = Request.Builder()
+            .url("${baseUrl}/telematics_commands")
+            .headers(baseHeaders)
+            .post(
+                requestBody(
+                    mapOf(
+                        "serial_number" to accessCertificate.gainerSerial.hex,
+                        "issuer" to certificate.issuer.hex,
+                        "data" to encryptedCommand.base64
+                    )
+                )
+            )
+            .build()
+
+        printRequest(request)
+
+        val call = client.newCall(request)
+        val response = call.await()
+
+        return tryParseResponse(response, HttpURLConnection.HTTP_OK) { body ->
+            val jsonResponse = jsonIg.parseToJsonElement(body) as JsonObject
+            val encryptedResponseCommand =
+                jsonResponse.jsonObject["response_data"]?.jsonPrimitive?.content
+            Response(Bytes(encryptedResponseCommand), null)
+        }
     }
 }